test: update after sync
