+++ frontend/src/tests/lib/components/accounts/CkBTCAccountsFooter.spec.ts
@@ -13,6 +13,7 @@ import {
   mockCkBTCMainAccount,
 } from "$tests/mocks/ckbtc-accounts.mock";
 import { mockTokens } from "$tests/mocks/tokens.mock";
+import { selectSegmentBTC } from "$tests/utils/accounts.test-utils";
 import { fireEvent } from "@testing-library/dom";
 import { render, waitFor } from "@testing-library/svelte";
 import { page } from "../../../../../__mocks__/$app/stores";
@@ -131,6 +132,8 @@ describe("CkBTCAccountsFooter", () => {
       await waitFor(() =>
         expect(container.querySelector("div.modal")).not.toBeNull()
       );
+
+      await selectSegmentBTC(container);
 
       const spy = jest.spyOn(services, "syncCkBTCAccounts");
